+++ statprly/mono_digits_recognizer/standards_provider/__init__.py
@@ -1,1 +1,3 @@
-from statprly.mono_digits_recognizer.standards_provider.provider import StandardsProvider
+from statprly.mono_digits_recognizer.standards_provider.provider import (
+    StandardsProvider,
+)
